requirements.txt
Cython==0.25.2
Pillow==4.0.0
argparse==1.2.1
cycler==0.10.0
dask==0.13.0
decorator==4.0.11
functools32==3.2.3-2
matplotlib==2.0.0
networkx==1.11
numpy==1.12.0
olefile==0.44
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
scikit-image==0.12.3
scipy==0.18.1
six==1.10.0
subprocess32==3.2.7
toolz==0.8.2
wsgiref==0.1.2
